Debug access to ChunkProviderServer#chunks
Detects async modifications and _any_ modification whilst iterating. Note this
includes the Iterator#remove method. If the Iterator#remove method is used,
then please notify me.

This also logs ALL removes/adds to a single log. This is such that in the case
that the map is faulty, we can reproduce the bug far more easier.